+++ e_ood/__init__.py
@@ -1,5 +1,5 @@
 """Convenience import of API symbols from e_ood modules"""
-__version__ = '0.1.40a1'
+__version__ = '0.2.0'
 # !__version__ must be set on the second line, for setup.py!
 
 from .analyze import Analyzer  # noqa